Refactor previews api to keep it simple, move implementation back to docconv.client
