+++ src/ploneintranet/api/previews.py
@@ -1,128 +1,7 @@
 # -*- coding: utf-8 -*-
-from plone import api
+''' Methods to generate and access preview images on content '''
 
-# from ploneintranet.async.celerytasks import generate_and_add_preview, app
-from collective.documentviewer.settings import Settings
-from collective.documentviewer.settings import GlobalSettings
-from collective.documentviewer.convert import DUMP_FILENAME
-from collective.documentviewer.convert import TEXT_REL_PATHNAME
-from collective.documentviewer import storage
-
-from collective.documentviewer.utils import allowedDocumentType
-from collective.documentviewer.utils import getPortal
-from collective.documentviewer.convert import Converter as DVConverter
-
-from zope.site.hooks import getSite
-
-from logging import getLogger
-
-log = getLogger(__name__)
-
-PREVIEW_URL = '++theme++ploneintranet.theme/generated/media/logo.svg'
-
-
-class Converter(DVConverter):
-    """ This class overrides the Converter class of collective.documentviewer
-        which forces its own layout on converted content objects. We need to
-        use our own layout, so this is deactivated.
-    """
-
-    def handle_layout(self):
-        """
-        Deactivate the layout setting part of documentviewer as we want our own
-        """
-        pass
-
-
-def _backward_map(scale):
-    """ Provide a mapping from plone scale names to collective.documentviewer
-    scale names for convenience. This assures that code can use both naming
-    conventions.
-
-    :param scale: The name of a scale
-    :type scale: A string
-    :return: The corresponding name of the scale as understood by c.dv
-    :rtype: string
-    """
-    if scale == 'thumb':
-        return 'small'
-    elif scale == 'preview':
-        return 'normal'
-    elif scale == 'large':
-        return 'large'
-    if scale not in ('large', 'normal', 'small'):
-        scale = 'large'  # default to a size that will never be pixeled
-
-    return scale
-
-
-def _get_dv_data(obj):
-    """ Access the collective.documentviewer settings of an object and
-    return metadata that can be used to retrieve or access a preview
-    image.
-
-    :param obj: The Plone content object that has a preview
-    :type obj: A Plone content object
-    :return: Metadata consisting of canonical_url, number of pages and
-    resource urls to preview images and thumbnails.
-    :rtype: mapping
-    """
-    site = getSite()
-    global_settings = GlobalSettings(site)
-    settings = Settings(obj)
-    portal_url = site.portal_url()
-
-    try:
-        canonical_url = obj.absolute_url()
-    except AttributeError:
-        canonical_url = ''  # XXX construct a url to an attachment
-
-    if not hasattr(obj, 'UID') or not settings.successfully_converted:
-        # Can't have previews on objects without a UID. Return a
-        # minimal datastructure
-        return {
-            'canonical_url': canonical_url + '/view',
-            'pages': settings.num_pages,
-            'resources': {
-                'page': {
-                    'image': fallback_image_url(obj),
-                    'text': ''
-                },
-                'pdf': canonical_url,
-                'thumbnail': fallback_image_url(obj),
-            }
-        }
-
-    resource_url = global_settings.override_base_resource_url
-    rel_url = storage.getResourceRelURL(gsettings=global_settings,
-                                        settings=settings)
-    if resource_url:
-        dvpdffiles = '%s/%s' % (resource_url.rstrip('/'),
-                                rel_url)
-    else:
-        dvpdffiles = '%s/%s' % (portal_url, rel_url)
-    dump_path = DUMP_FILENAME.rsplit('.', 1)[0]
-
-    image_format = settings.pdf_image_format or \
-        global_settings.pdf_image_format
-
-    return {
-        'canonical_url': canonical_url + '/view',
-        'pages': settings.num_pages,
-        'resources': {
-            'page': {
-                'image': '%s/{size}/%s_{page}.%s' % (
-                    dvpdffiles, dump_path,
-                    image_format),
-                'text': '%s/%s/%s_{page}.txt' % (
-                    dvpdffiles, TEXT_REL_PATHNAME, dump_path)
-            },
-            'pdf': canonical_url,
-            'thumbnail': '%s/small/%s_1.%s' % (
-                dvpdffiles, dump_path,
-                image_format),
-        }
-    }
+from ploneintranet.docconv.client import previews
 
 
 def get(obj, scale='normal'):
@@ -135,15 +14,7 @@ def get(obj, scale='normal'):
     :return: The preview images
     :rtype: list
     """
-    previews = []
-    settings = Settings(obj)
-    if settings.blob_files:
-        ext = settings.pdf_image_format
-        scale = _backward_map(scale)
-        for i in range(settings.num_pages):
-            preview = '%s/dump_%s.%s' % (scale, str(i + 1), ext)
-            previews.append(settings.blob_files[preview])
-    return previews
+    return previews.get(obj, scale)
 
 
 def has_previews(obj):
@@ -154,9 +25,7 @@ def has_previews(obj):
     :return: True if there are previews. False otherwise.
     :rtype: boolean
     """
-    if get(obj):
-        return True
-    return False
+    return previews.has_previews(obj)
 
 
 def get_preview_urls(obj, scale='normal'):
@@ -170,17 +39,7 @@ def get_preview_urls(obj, scale='normal'):
     :return: List of preview image absolute URLs
     :rtype: list
     """
-    dv_data = _get_dv_data(obj)
-    number_of_previews = dv_data['pages']
-
-    # If there aren't any previews, return the placeholder url
-    if number_of_previews < 1:
-        return [fallback_image_url()]
-    scale = _backward_map(scale)
-    return [dv_data['resources']['page']['image'].format(size=scale,
-                                                         page=page)
-            for page in range(1, number_of_previews + 1)
-            ]
+    return previews.get_preview_urls(obj, scale)
 
 
 def fallback_image(obj):
@@ -191,7 +50,7 @@ def fallback_image(obj):
     :return: An image object
     :rtype: ZODB.blob.Blob
     """
-    return api.portal.get().restrictedTraverse(PREVIEW_URL)
+    return previews.fallback_image(obj)
 
 
 def fallback_image_url(obj):
@@ -202,7 +61,7 @@ def fallback_image_url(obj):
     :return: An image object url
     :rtype: string
     """
-    return '{0}/{1}'.format(api.portal.get().absolute_url(), PREVIEW_URL)
+    return previews.fallback_image_url(obj)
 
 
 def get_thumbnail(obj):
@@ -213,10 +72,7 @@ def get_thumbnail(obj):
     :return: The preview thumbnail
     :rtype: ImageScaling
     """
-    previews = get(obj, scale='small')
-    if len(previews) < 1:
-        return fallback_image(obj)
-    return previews[0]
+    return previews.get_thumbnail(obj)
 
 
 def get_thumbnail_url(obj, relative=False):
@@ -227,13 +83,7 @@ def get_thumbnail_url(obj, relative=False):
     :return: The absolute URL to the thumbnail image
     :rtype: str
     """
-    dv_data = _get_dv_data(obj)
-    thumbnail_url = dv_data['resources']['thumbnail']
-    if relative:
-        portal = api.portal.get()
-        thumbnail_url = thumbnail_url.replace(portal.absolute_url(),
-                                              portal.absolute_url(1))
-    return thumbnail_url
+    return previews.get_thumbnail_url(obj, relative)
 
 
 def has_pdf(obj):
@@ -268,15 +118,4 @@ def generate_previews(obj, event=None):
     :return: Does not return anything.
     :rtype: None
     """
-    site = getPortal(obj)
-    gsettings = GlobalSettings(site)
-
-    if not allowedDocumentType(obj, gsettings.auto_layout_file_types):
-        return
-
-    if gsettings.auto_convert:
-        # ASYNC HERE
-        converter = Converter(obj)
-        if not converter.can_convert:
-            return
-        converter()
+    previews.generate_previews(obj, event)
